requirements.txt
absl-py==0.1.10
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
cycler==0.10.0
docopt==0.6.2
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
html5lib==0.9999999
Markdown==2.6.11
matplotlib==2.1.2
mock==2.0.0
numpy==1.14.1
pandas==0.22.0
pbr==3.1.1
pnr==0.0.1
protobuf==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
six==1.11.0
subprocess32==3.2.7
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0
tqdm==4.19.5
Werkzeug==0.14.1
